+++ internal/db/migrations/2025-06-10.3-create_indexer_table_operations.sql
@@ -2,7 +2,7 @@
 
 -- Table: operations
 CREATE TABLE operations (
-    id INTEGER PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     tx_hash TEXT NOT NULL REFERENCES transactions(hash),
     operation_type TEXT NOT NULL,
     operation_xdr TEXT,
@@ -16,7 +16,7 @@ CREATE INDEX idx_operations_ledger_created_at ON operations(ledger_created_at);
 
 -- Table: operations_accounts
 CREATE TABLE operations_accounts (
-    operation_id INTEGER NOT NULL REFERENCES operations(id) ON DELETE CASCADE,
+    operation_id BIGINT NOT NULL REFERENCES operations(id) ON DELETE CASCADE,
     account_id TEXT NOT NULL REFERENCES accounts(stellar_address) ON DELETE CASCADE,
     created_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
     PRIMARY KEY (account_id, operation_id)
